Remove weather lookup example
Ref #4769
