+++ java/src/main/java/com/google/oak/client/oak_functions_client/README.md
@@ -1,0 +1,9 @@
+# Java Oak Functions Client
+
+This in the Java version of the Oak Functions Client.
+
+Build and run the client with the following command:
+
+```bash
+bazel run //java/src/main/java/com/google/oak/client/oak_functions_client -- http://localhost:8080
+```
