--- java/src/main/java/com/google/oak/client/weather_lookup_client/README.md
@@ -1,9 +1,0 @@
-# Java gRPC Attestation Client
-
-This in the Java version of the gRPC Attestation client.
-
-Build and run the client with the following command:
-
-```bash
-bazel run //java/src/main/java/com/google/oak/client/weather_lookup_client -- http://localhost:8080
-```
