--- oak_functions/examples/weather_lookup/module/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "weather_lookup"
-version = "0.1.0"
-authors = ["Razieh Behjati <razieh@google.com>"]
-edition = "2021"
-license = "Apache-2.0"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-location_utils = { workspace = true }
-oak_functions_sdk = { workspace = true }
-serde = { version = "*", features = ["derive"] }
-serde_json = "*"
-
-[dev-dependencies]
-lookup_data_generator = { path = "../../../lookup_data_generator" }
-oak_functions_abi = { workspace = true }
-maplit = "*"
-rand = "*"
-oak_functions_test_utils = { workspace = true }
-tokio = { version = "*", features = [
-  "fs",
-  "macros",
-  "sync",
-  "rt-multi-thread"
-] }
-tonic = "*"
-xtask = { workspace = true }
-log = "*"
