--- oak_functions/load_test/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "oak_functions_load_test"
-version = "0.1.0"
-authors = ["Tiziano Santoro <tzn@google.com>"]
-edition = "2021"
-license = "Apache-2.0"
-
-[dependencies]
-anyhow = "*"
-bencher = "*"
-oak_client = { workspace = true }
-oak_functions_client = { workspace = true }
-rand = "*"
-tokio = { workspace = true, features = [
-  "fs",
-  "macros",
-  "sync",
-  "rt-multi-thread"
-] }
