--- oak_functions/location_utils/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "location_utils"
-version = "0.1.0"
-authors = ["Conrad Grobler <grobler@google.com>"]
-edition = "2021"
-license = "Apache-2.0"
-
-[dependencies]
-anyhow = "*"
-s2 = { version = "*", default-features = false }
