--- oak_functions/lookup_data_checker/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "lookup_data_checker"
-version = "0.1.0"
-authors = ["Ivan Petrov <ivanpetrov@google.com>"]
-edition = "2021"
-license = "Apache-2.0"
-
-[dependencies]
-anyhow = "*"
-bytes = "*"
-clap = { version = "*", features = ["derive"] }
-env_logger = "*"
-location_utils = { workspace = true }
-log = "*"
-oak_proto_rust = { workspace = true }
-prost = { workspace = true }
