+++ oak_functions/lookup_data_generator/README.md
@@ -3,7 +3,3 @@
 The binary in this directory generates and serializes a list of entries with
 configurable key and value size, to be used to benchmark and test the lookup
 functionality of the Oak Functions platform.
-
-See
-[`generate_and_upload_lookup_data_weather`](/scripts/generate_and_upload_lookup_data_weather)
-for an example use.
